Rename the stats accordion component and added secondary titles
